requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0
drf-spectacular>=0.26.0
django-filter>=23.0
mysqlclient>=2.1.0